refactor: remove frontend code
