--- next/web/src/App/Admin/Tickets/Filter/FilterMenu/index.module.css
@@ -1,15 +1,0 @@
-.menu {
-  box-shadow: 4px 0 13px 0 rgba(74, 74, 74, 0.3), 30px 0 40px 0 rgba(0, 0, 0, 0.11);
-}
-
-.saveAs {
-  position: relative;
-
-  &::after {
-    content: '..';
-    position: absolute;
-    right: 0;
-    bottom: 0;
-    font-size: 16px;
-  }
-}
