+++ next/web/src/App/Admin/Tickets/Filter/index.ts
@@ -1,2 +1,2 @@
 export * from './FilterForm';
-export { FilterMenuTriggerPortal as FilterMenu } from './FilterMenu';
+export * from './useTicketFilter';
